Added operation text for swagger and changed api endpoints to a good name.

+++ src/main/java/com/backend/chatapplicatie/controllers/UserController.java
@@ -1,18 +1,21 @@
 package com.backend.chatapplicatie.controllers;
 
 import com.backend.chatapplicatie.models.User;
+import com.backend.chatapplicatie.payload.response.MessageResponse;
 import com.backend.chatapplicatie.security.services.UserService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
 @CrossOrigin(origins = "http://localhost:3000")
-@RequestMapping("/users")
-@PreAuthorize("hasRole('ROLE_ADMIN')")
+@RequestMapping("/usermanagement")
+@Tag(name = "CRUD")
 public class UserController {
 
     private final UserService userService;
@@ -21,21 +24,20 @@ public class UserController {
     public UserController(UserService userService) {
         this.userService = userService;
     }
-    /*Get all users */
 
-    @GetMapping("/getall")
+    @GetMapping("/users")
+    @Operation(summary = "Alle gebruikers ophalen", description = "Haal een lijst op van alle gebruikers.")
     public ResponseEntity<List<User>> getAllUsers() {
         List<User> users = userService.getAllUsers();
         return ResponseEntity.ok(users);
     }
-    /*Delete by ID */
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
-        if (userService.userExists(id)) {
-            userService.deleteUser(id);
-            return ResponseEntity.noContent().build();
+    @Operation(summary = "Gebruiker verwijderen op basis van ID", description = "Verwijder een gebruiker op basis van hun ID.")
+    public ResponseEntity<MessageResponse> deleteUser(@PathVariable Long id) {
+        if (userService.userExists(id)) { userService.deleteUser(id);
+            return ResponseEntity.ok(new MessageResponse("Gebruiker succesvol verwijderd"));
         } else {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new MessageResponse("Geen gebruiker gevonden"));
         }
     }
 }
